# backend/openapi-puzzle-games.yaml
openapi: 3.0.3
info:
  title: 益智游戏平台 API
  version: puzzle-api-v1
  description: 每日数独、成语填字、记忆翻牌、字阵寻踪和游戏大厅接口。成功直接返回 JSON，失败返回 Error。
servers:
  - url: /api
security:
  - {}
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /sudoku/save:
    post:
      summary: 保存数独云端进度
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SudokuSaveRequest'
      responses:
        '200':
          $ref: '#/components/responses/Saved'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /sudoku/hint:
    post:
      summary: 获取一个数独提示
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SudokuHintRequest'
      responses:
        '200':
          description: 一个尚未正确填写的格子
          content:
            application/json:
              schema:
                type: object
                required: [index, row, column, value, hints_used, remaining_hints]
                properties:
                  index: {type: integer, example: 10}
                  row: {type: integer, example: 1}
                  column: {type: integer, example: 1}
                  value: {type: integer, example: 7}
                  hints_used: {type: integer, example: 1}
                  remaining_hints: {type: integer, example: 2}
        '403':
          $ref: '#/components/responses/Forbidden'
  /sudoku/submit:
    post:
      summary: 校验并完成数独
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SudokuSubmitRequest'
      responses:
        '200':
          description: 正确完成或幂等返回原结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Completion'
        '422':
          description: 答案不完整或不正确，不返回解答
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncorrectCells'
...
  /idiom/save:
    post:
      summary: 保存成语填字云端进度
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdiomSaveRequest'
      responses:
        '200':
          $ref: '#/components/responses/Saved'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /idiom/submit:
    post:
      summary: 校验并完成成语题
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdiomSubmitRequest'
      responses:
        '200':
          description: 正确完成或幂等返回原结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Completion'
        '422':
          description: 有未填或错误文字
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IncorrectCells'
...
  /memory/save:
    post:
      summary: 保存翻牌云端进度
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryProgressRequest'
      responses:
        '200':
          $ref: '#/components/responses/Saved'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /memory/submit:
    post:
      summary: 保存翻牌完成结果
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryProgressRequest'
      responses:
        '200':
          description: 正确完成或幂等返回原结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Completion'
        '422':
          description: 仍有未匹配牌面
          content:
            application/json:
              schema:
                type: object
                properties:
                  correct: {type: boolean, example: false}
                  status: {type: string, example: incorrect}
                  unmatched_count: {type: integer, example: 2}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: 网页端 Flask 签名会话 Cookie；请求需允许携带 credentials
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token
...
  responses:
    Saved:
      description: 云端进度已保存
      content:
        application/json:
          schema:
            type: object
            required: [saved, run_id, updated_at]
            properties:
              saved: {type: boolean, example: true}
              run_id: {type: string, example: run_f9c4a03e}
              updated_at: {type: string, format: date-time, example: '2026-07-16T09:32:10Z'}
...
    Unauthorized:
      description: 未登录或 Token 失效
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
    Forbidden:
      description: 未解锁或提示次数已用完
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    Conflict:
      description: 运行记录状态冲突
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error: {type: string, example: 题目不存在}
        code: {type: string, example: PUZZLE_NOT_FOUND}
...
    SudokuState:
      type: object
      properties:
        grid: {type: string, minLength: 81, maxLength: 81}
        notes:
          type: object
          additionalProperties:
            type: array
            maxItems: 9
            items: {type: integer, minimum: 1, maximum: 9}
        elapsed_seconds: {type: integer, minimum: 0}
        hints_used: {type: integer, minimum: 0, maximum: 3}
        mistakes: {type: integer, minimum: 0}
    SudokuSaveRequest:
      allOf:
        - $ref: '#/components/schemas/SudokuState'
        - type: object
          required: [run_id, puzzle_id, grid, elapsed_seconds, mistakes]
          properties:
            run_id: {type: string}
            puzzle_id: {type: string}
    SudokuHintRequest:
      type: object
      required: [puzzle_id, grid]
      properties:
        run_id: {type: string, nullable: true}
        puzzle_id: {type: string}
        grid: {type: string, minLength: 81, maxLength: 81}
    SudokuSubmitRequest:
      allOf:
        - $ref: '#/components/schemas/SudokuHintRequest'
        - type: object
          required: [elapsed_seconds, mistakes]
          properties:
            elapsed_seconds: {type: integer, minimum: 0}
            mistakes: {type: integer, minimum: 0}
            hints_used: {type: integer, minimum: 0, description: 游客展示用；登录用户以服务端累计值为准}
...
    IdiomSaveRequest:
      type: object
      required: [run_id, puzzle_id, grid, elapsed_seconds, mistakes]
      properties:
        run_id: {type: string}
        puzzle_id: {type: string}
        grid: {type: array, maxItems: 32, items: {type: string, maxLength: 1}}
        elapsed_seconds: {type: integer, minimum: 0}
        mistakes: {type: integer, minimum: 0}
    IdiomHintRequest:
      type: object
      required: [puzzle_id, grid]
      properties:
        run_id: {type: string, nullable: true}
        puzzle_id: {type: string}
        grid: {type: array, maxItems: 32, items: {type: string, maxLength: 1}}
        entry_id: {type: string, nullable: true}
    IdiomSubmitRequest:
      allOf:
        - $ref: '#/components/schemas/IdiomHintRequest'
        - type: object
          required: [elapsed_seconds, mistakes]
          properties:
            elapsed_seconds: {type: integer, minimum: 0}
            mistakes: {type: integer, minimum: 0}
            hints_used: {type: integer, minimum: 0}
...
    MemoryProgressRequest:
      type: object
      required: [board_id, matched_positions, moves, elapsed_seconds]
      properties:
        run_id: {type: string, nullable: true}
        board_id: {type: string, maxLength: 120}
        matched_positions:
          type: array
          maxItems: 30
          uniqueItems: true
          items: {type: integer, minimum: 0}
        moves: {type: integer, minimum: 0}
        elapsed_seconds: {type: integer, minimum: 0}
...
    IncorrectCells:
      type: object
      required: [correct, status, invalid_cells]
      properties:
        correct: {type: boolean, example: false}
        status: {type: string, example: incorrect}
        invalid_cells: {type: array, items: {type: integer}}
    Completion:
      type: object
      required: [correct, status, result]
      properties:
        correct: {type: boolean, example: true}
        status: {type: string, example: completed}
        result:
          type: object
          properties:
            score: {type: integer, example: 920}
            stars: {type: integer, minimum: 1, maximum: 3}
            elapsed_seconds: {type: integer}
            mistakes: {type: integer}
            hints_used: {type: integer}
            moves: {type: integer}
            earned_stars: {type: integer}
            total_stars: {type: integer}
            next_level_id: {type: string, nullable: true}
            is_new_best: {type: boolean}